Add github action for docs (#95)
* add github action for docs

* patch few issues

* fix paths

* correct module path

* Configures GitHub Pages deployment workflow

Updates the documentation deployment workflow to use the official GitHub Pages actions.

This change replaces the previous `peaceiris/actions-gh-pages` action with `actions/configure-pages` and `actions/deploy-pages` for better integration and control over the deployment process.

Additionally, it adds write permissions to the GitHub token for contents, pages, and id-token to allow for successful deployment.

* Adds environment for GitHub Pages deployment

Configures the environment section in the workflow to
define the GitHub Pages deployment target. This enables
GitHub to track deployments to the github-pages
environment and provides a URL.

* feat: install package in editable mode for sphinx docs

* feat: enable Pages in workflow configuration

* Trigger build

* Trigger again, with deploy from action

* Test doc string

---------

Co-authored-by: Sebastian Jiro Schlecht <sebastianjiroschlecht@gmail.com>